Fix order of converted lists
For performance reasons, the code conses new entries the the head of
a partial list during conversion. Since we ran from the first to the
last entry in a list, that technique caused the resulting list to be
inverted. This fixes the issue by starting at the end and running to
the top.

Signed-off-by: Frank Terbeck <ft@bewatermyfriend.org>