value SCM: Add a back-reference to the parent result
With synchronous clients, you extract the `xmmsv_t' container from
the `xmmsc_result_t' container, and the xmmv_t is freed by the XMMS2
client library together with its parent xmmsc_result_t.

Without this back-reference we cannot keep the parent referenced while
working with the value. This occationally caused segfaults.

With asynchronous clients, there is no parent xmmsc_result_t. But the
code should just work, because we just keep a reference to SCM_EOL,
which will always be available in a lisp system.

Signed-off-by: Frank Terbeck <ft@bewatermyfriend.org>