Initial commit
This marks the start of a new line of development with xmms2-guile. This
branch sets out to implement a native implementation of the xmms2
protocol as opposed to wrapping the C based client library from the
xmms2 project itself.

Wrapping the C library was the previous approach of this project and its
code is still available in the “wrap-c-lib” branch.

Signed-off-by: Frank Terbeck <ft@bewatermyfriend.org>