Add a tiny little bit of documentation about the protocol
XMMS2's (binary) protocol is pretty much undocumented. The document
added in this commit adds a scratch-pad like piece of documentation
describing my current understanding of what's going on in the protocol.

The idea is to put it into a more polished form when I'm confident that
its contents are more or less correct and complete.

Signed-off-by: Frank Terbeck <ft@bewatermyfriend.org>