compile: Make conditional work in POSIX shells
I wrote this code on a machine that had bash being /bin/sh, now
with dash as /bin/sh the conditional fails if the $output does
not exist yet.

I can't be arsed to look at the POSIX spec right now if that is
indeed the correct behaviour. This makes it work in any case.

Signed-off-by: Frank Terbeck <ft@bewatermyfriend.org>