Add (xmms2 io) module
This module contains an abstraction for xmms2 server connections (tcp
and unix sockets) and assorted utilities to perform connects,
disconnects, sending and receiving with such connections.

This is the bare metal. The rest of the libraries job is to generate
packets for it to send, have disectors for received data and to
implement higher level abstractions for application usage.

Signed-off-by: Frank Terbeck <ft@bewatermyfriend.org>