Add IPC types and target language macro
In order to make code generation easier, the define-ipc-packet-generator
macro is provided as a target language. This allows code generation like
this:

  (define-ipc-packet-generator make-hello public OBJECT-MAIN CMD-HELLO
    "Say 'Hello' to the server"
    (integer protocol-version) (string client-name))

Which will generate a function like this:

  (define-public (make-hello protocol-version client-name)
    ...)

The function will test the type of the parameter values passed to it. If
even one of them is false, the function will thrown ‘type-error’.

This target language will allow us to define the actual code that will
be generated from within the library, rather than at genipc time, which
I hope will only happen when absolutely required.

Signed-off-by: Frank Terbeck <ft@bewatermyfriend.org>