genipc: Actually generate scheme/xmms2/ipc/*
This only generates packet generator function for methods against the
define-ipc-packet-generator language. The documentation generated can
still be improved, as the TODO comment states.

Signals and broadcasts packet generator functions will be pretty simple
after this. In addition, the introspection structures need to be
generated as well.

This will also show, that a number of payload generators are still
missing, that are required to implement all of the methods so far:

  - make-list-payload (should be used for method argument lists, too)
  - make-binary-payload
  - make-collection-payload
  - make-dictionary-payload

On top of that, the define-ipc-packet-generator type-checker needs
support for these.

On the plus side, simple methods like start, stop and pause from the
(xmms2 ipc playlist) module do work fine; which means the library can
now control an XMMS2 server. Albeit in limited manner thus far.

Signed-off-by: Frank Terbeck <ft@bewatermyfriend.org>