Don't kill the compiler if a type is unknown
This should be fatal in the future, but it makes development harder than
it needs to be right now. As soon as the library is more complete, this
needs to die as loud as possible, since it would clearly show missing
features in the library that will throw errors at run-time for the user.

Signed-off-by: Frank Terbeck <ft@bewatermyfriend.org>