Implement floating point payload
This was a little harder than expected, because I had to implement
‘frexp’ first. Then I was comparing the values produced by my imple-
mentation to those of a reference implemenation I took right out of
xmms2's core… and as it turns out there are differences. Fairly small
ones, but still. So I added a few tests that make sure that the payload
this library produces are at least as close to the real value as the
generator from the core library.

The test program I used was this:

    #include <math.h>
    #include <float.h>
    #include <stdint.h>
    #include <stdio.h>
    #include <stdlib.h>

    void
    internal_put_on_bb_float(float v)
    {
        int64_t mant_int;
        int32_t expon;
        float mant;

        mant = frexpf(v, &expon);
        if (mant > 0.0f)
            mant_int = mant * INT32_MAX;
        else
            mant_int = mant * fabsf(INT32_MIN);

        printf(" (test-float-payload %e #vu8(0 0 0 9", v);
        for (int i = 3; i >= 0; --i)
            printf(" %u", (mant_int >> (8*i)) & 0xff);
        for (int i = 3; i >= 0; --i)
            printf(" %u", (expon >> (8*i)) & 0xff);
        printf("))\n");
    }

    int
    main(int argc, char *argv[])
    {
        double x, r;
        int exp;
        x = strtod(argv[1], NULL);
        r = frexp(x, &exp);
        internal_put_on_bb_float(x);
        return EXIT_SUCCESS;
    }

Signed-off-by: Frank Terbeck <ft@bewatermyfriend.org>