tests: Add a back and forth test macro
Takes a generator, a disector and a value, passes the value into the
generator and turns the payload from that back into a value and checks
if the resulting value matches the input value.

The macro also takes the name of a test-tap macro that is supposed to be
used to perform the actual test. Any excess arguments are passed to the
test verbatim. This way you can do:

Exact test for integers:

    (perform-payload-><-test make-int64-payload
                             payload->int64
                             123 pass-if-=)

Inexact test for floating values:

    (perform-payload-><-test make-float-payload
                             payload->float
                             (exact->inexact (/ 2 3)
                             pass-if-~= 1e-6)

Signed-off-by: Frank Terbeck <ft@bewatermyfriend.org>