Introduce a dedicated dictionary type
I think it's still the best approach to use association lists as a represen-
tation for dictionaries in Scheme. But it's kind of hard to tell empty lists
and empty dictionaries apart, which is a problem anytime a piece of data flows
through make-value-payload.

This approaches the problem list this:

  - The old dictionary? predicate is now association-list?
  - There is a new record type <dictionary> that represents dictionaries.
  - It comes with its own dictionary? predicate.
  - This allows empty dictionaries to be recognised as such.

All functions that processed dictionaries before still happily take data that
satisfies the association-list? predicate, but now in addition support data the
satisfies the dictionary? predicate.

Signed-off-by: Frank Terbeck <ft@bewatermyfriend.org>