collection DSL: Allow arbitrarily complex expressions in arguments
To let users express collection is a convenient manner, the DSL supports
automatically-stringified arguments to its operators:

  (collection (artist = Slayer))

This is nice, but what to do it you want to create a collection that matches an
artist of whom you got the name of in a variable? This commit allows the use of
parenthesized expressions to force the language to evaluate arguments:

  (let ((name "Slayer"))
    (collection (artist = (name))))

This allows for arbitrarily complex expressions to be used in these positions.

Signed-off-by: Frank Terbeck <ft@bewatermyfriend.org>