payload: Do not use collection-fold to recurse on collections
The systems works recursively already: make-collection-payload will call
make-list-payload for the operand collection list and that will find items of
type <collection> in that list and therefore call make-collection-payload again
to deal with them.

Using collection-fold caused excess data to be generated! The first bytes where
the correct serialisation of the collection and after that, data was added that
are the operand collections serialised again.

Because the deserializers could happily deserialise the collection from the
first few bytes, everything looked good in the test suite, where we feed the
result of our serialiser into our deserialiser.

In general, this if good, because we want to be able to extract a collection
from anywhere in a byte stream. But in case of these tests it is less than
optimal.

In order to make the test suite stricter, the tests should check if the
deserializers consumed all data that was produced by a serialiser, because in
these particular test cases, they MUST.

Signed-off-by: Frank Terbeck <ft@bewatermyfriend.org>