Add a documentation gathering utility module
With this, it is possible to gather documentation for a given module. It is
possible to register additional documentation for a module

  (xmms2 foobar)

in the

  (documentation more foobar)

namespace, like is done with (documentation more io). This is useful to add
documentation for pieces of API that can't have documentation attached to them
natively in Guile. Like constant values as in

  (define *foobar* 1234)

The plan is to use this to generate most of the API documentation on the fly
from source for inclusion in the markdown source of the reference manual.

Signed-off-by: Frank Terbeck <ft@bewatermyfriend.org>